+++ HISTORY.md
@@ -11,6 +11,9 @@
 
   * Add `ClassProbabilities()` member to `DecisionTree` so that the internal
     details of trees can be more easily inspected (#3511).
+  
+  * Bipolar sigmoid activation function added and invertible functions 
+    fixed (#3506).
 
 ### mlpack 4.2.0
 ###### 2023-06-14